Revert "Get all inventory items"
