+++ .gitignore
@@ -39,5 +39,3 @@ node_modules/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
-.eslintcache
