+++ client/.eslintcache
@@ -1,2 +1,1 @@
 [{"C:\\Users\\HARGUL\\Desktop\\Brainstation\\Monutments_Men\\mm-inStock\\client\\src\\components\\Header.jsx":"1"},{"size":299,"mtime":1606268932247,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"7foe2k","C:\\Users\\HARGUL\\Desktop\\Brainstation\\Monutments_Men\\mm-inStock\\client\\src\\components\\Header.jsx",[]]
-
