--- client/src/components/WarehouseDetails.jsx
@@ -1,5 +1,0 @@
-function WarehouseDetails() {
-  return <h1>Warehouse Details</h1>
-}
-
-export default WarehouseDetails
