+++ client/src/style/App.css
@@ -4,26 +4,6 @@
   border: 1px dashed grey;
 }
 
-body {
-  margin: 0;
-  padding: 0;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-li {
-  list-style-type: none;
-}
-
-.card-look {
-  max-width: 21.438rem;
-  margin-top: -4rem;
-  background-color: white;
-}
-
 .header {
   background-color: #232940;
   height: 12.5rem;
