+++ client/src/style/partials/_bones.scss
@@ -2,21 +2,3 @@
   box-sizing: border-box;
   border: 1px dashed grey;
 }
-
-body {
-  margin: 0;
-  padding: 0;
-}
-a {
-  color: inherit;
-  text-decoration: none;
-}
-li {
-  list-style-type: none;
-}
-
-.card-look {
-  max-width: 21.438rem;
-  margin-top: -4rem;
-  background-color: white;
-}
