+++ server/package.json
@@ -11,6 +11,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "uuid": "^8.3.1"
@@ -18,4 +19,5 @@
   "devDependencies": {
     "nodemon": "^2.0.6"
   }
+
 }
